feat: Add scenario input to checkpointer
