+++ crates/checkpointer/Cargo.toml
@@ -16,6 +16,7 @@ chrono.workspace = true
 clap = { workspace = true, features = ["derive"] }
 data-generation = { path = "../data-generation" }
 etl = { path = "../etl", features = ["duckdb"] }
+test-framework = { path = "../test-framework" }
 object_store = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
 serde_json.workspace = true
